feat: Define OffsetRecordRepository interface
- This OffsetRecordRepository getting the file key offset from the OffsetManager instead of Kafka.
